+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Prefix the change with one of these keywords:
 - _Security_: in case of vulnerabilities.
 
 ## Unreleased
+- Added Hidden component
 
 ## [0.19.1]
 - Changed: AmsterdamLogo is now the default logo in the Header